Allow launching UI from CLI
